+++ CardGame/CardGame/PokerGamePlay.swift
@@ -3,7 +3,7 @@
 //  CardGame
 //
 //  Created by TaeHyeonLee on 2017. 12. 4..
-//  Copyright © 2017년 JK. All rights reserved.
+//  Copyright © 2017 ChocOZerO. All rights reserved.
 //
 
 import Foundation
